Sync

+++ examples/grad_compare_testnet.py
@@ -53,7 +53,7 @@ def _parse_dtype(value: str) -> torch.dtype:
 
 def main() -> None:
     parser = argparse.ArgumentParser(description="Compare streaming vs non-streaming backward gradients for ResNet18.")
-    parser.add_argument("--dtype", default="float64", help="float16, float32, or float64")
+    parser.add_argument("--dtype", default="float32", help="float16, float32, or float64")
     parser.add_argument("--tile-size", type=int, default=3200)
     parser.add_argument("--input-size", type=int, default=4800)
     args = parser.parse_args()
